+++ setup.cfg
@@ -1,15 +1,15 @@
 [metadata]
-name = stactools-ephemeral
-version = attr: stactools.ephemeral.__version__
+name = stactools-drcog-lulc
+version = attr: stactools.drcog_lulc.__version__
 description = PROVIDE DESCRIPTION HERE
 long_description = file: README.md
 long_description_content_type = text/markdown
 author = stac-utils
 author_email = stac@radiant.earth
-url = https://github.com/stactools-packages/ephemeral
+url = https://github.com/stactools-packages/drcog-lulc
 project_urls =
-    Documentation = https://stactools-ephemeral.readthedocs.io/en/latest/
-    Issues = https://github.com/stactools-packages/ephemeral/issues
+    Documentation = https://stactools-drcog-lulc.readthedocs.io/en/latest/
+    Issues = https://github.com/stactools-packages/drcog-lulc/issues
 keywords =
     stactools
     pystac
