+++ src/stactools/drcog_lulc/__init__.py
@@ -1,0 +1,15 @@
+import stactools.core
+
+from stactools.drcog_lulc.stac import create_collection, create_item
+
+__all__ = ['create_collection', 'create_item']
+
+stactools.core.use_fsspec()
+
+
+def register_plugin(registry):
+    from stactools.drcog_lulc import commands
+    registry.register_subcommand(commands.create_drcoglulc_command)
+
+
+__version__ = "0.1.0"
