+++ tests/test_module.py
@@ -1,9 +1,9 @@
 import unittest
 
-import stactools.ephemeral
+import stactools.drcog_lulc
 
 
 class TestModule(unittest.TestCase):
 
     def test_version(self):
-        self.assertIsNotNone(stactools.ephemeral.__version__)
+        self.assertIsNotNone(stactools.drcog_lulc.__version__)
